Fix compatibility issue with Arduino Due
The accel_t_gyro_union doesn't work properly with the 32 bit Cortex processor. Changing int to short fixes the problem.

+++ imu/gy_521_send_serial.ino
@@ -652,13 +652,13 @@ typedef union accel_t_gyro_union
   } reg;
   struct 
   {
-    int x_accel;
-    int y_accel;
-    int z_accel;
-    int temperature;
-    int x_gyro;
-    int y_gyro;
-    int z_gyro;
+    short x_accel;
+    short y_accel;
+    short z_accel;
+    short temperature;
+    short x_gyro;
+    short y_gyro;
+    short z_gyro;
   } value;
 };
